PlotPlants

Set global font

The PyPlot package of Julia uses sans font by default, by my favorite font is serif type. Also, some fonts do not have normal greek letters in their library, like the Times only have italic greek letters. To change the default font to serif and use normal greek letters, the PlotPlants package provides a use_serif_tex function, which use LaTeX amsmath and upgreek packages.

PlotPlants.use_serif_tex — Function

use_serif_tex()

Set default font to serif and use latex wrapper for texts

Symbols and units

To facilitate the use of LaTeX packages, functions latex_symbol and latex_unit are provided to rapidly generate equation symbols and units. By default,

if the number of letters is one, the letter will be set to italic, otherwise, the letters will be normal (not italic);

if the letter is greek, the letter will be normal;

all sub- and super-scripts are normal;

PlotPlants.latex_symbol — Function

latex_symbol(mid::String; sub::String, sup::String, presub::String, presup::String)

Return the latex symbol string, given

mid Center symbol, italic only when length>1

sub Optional: subscript after the mid

sup Optional: supscript after the mid

presub Optional: subscript before the mid

presup Optional: supscript before the mid

PlotPlants.latex_unit — Function

latex_unit(id::String; paren::Bool)

Return the unit, given

id Identifier of the unit (if not in libary, use id as unit)

paren Optional: if true, include parenthesis

squared Optional: if true, use [], otherwise, use ()

Create canvas

An customized funtion create_canvas is provided to genrate figure and axes from a few keywords. For example,

create_canvas(fig_id) returns a figure with one axis;

create_canvas(fig_id, nrow=2, ncol=2) returns a figure will four subplots;

create_canvas(fig_id, nrow=2, ncol=2, ax_ind=[1,2,4]) returns a figure with three subplot (the left bottom panel is empty);

PlotPlants.create_canvas — Function

create_canvas(id; ncol::Int, nrow::Int, ax_ind::Array{Int,1}, figsize::Tuple{Number,Number}, dpi::Number)

Create a canvas, given

id ID of the figure

ncol Number of columns in the figure

nrow Number of rows in the figure

ax_ind Given indicies of the subplots in the figure

figsize Given canvas size

dpi Given pixels per inch

Shapes

A few customized shapes can be added to the canvas very conveniently, for example, stoma and ellipse. These shortcut functions include

PlotPlants.plot_ellipse — Function

plot_ellipse(ax, xy::Tuple{Number,Number}; width::Number, height::Number, angle::Number, color::String, edgecolor::String, facecolor::String, alpha::Number = 0.5)

Plot an ellipse on axis, given

ax Axis to plot on

xy Center of the ellipse

width Width of the ellipse

height Height of the ellipse

angle Rotation angle of the ellipse

color Color of the ellipse

edgecolor Edgecolor of the ellipse

facecolor Face color of the ellipse

alpha Transparency of the ellipse

PlotPlants.plot_stoma — Function

plot_stoma(ax, xy::Tuple{Number,Number}; width::Number, height::Number, stoma::Number, angle::Number)

Plot a stoma on the axis, given

ax Axis to plot on

xy Center of the stoma

width Width of the stoma

height Height of the stoma

stoma Stomatal pore width ratio

angle Rotation angle of the stoma

Add titles

Function set_titles! will add panel titles like (a) to all the panels, and if labels options is given, the labels will be added after the letter (e.g., (a) label).

PlotPlants.set_titles! — Function

set_titles!(axes::Array; labels::Array{String,1}, paren::Bool, capital::Bool, fontsize::Number, loc::String)

Set titles for the axes, given

axes An array of axes

labels Optional: labels after the panel title, e.g., (a) label

paren Optional: if true, use format like (a)

capital Optional: if true, use capital letters like (A)

fontsize Optional: fontsize of the title

loc Optional: location of the title

Add regression

To quickly plot the regression lines, users may use

PlotPlants.plot_line_regress — Function

plot_line_regress(ax, list_x::Array, list_y::Array; intercept::Bool, interval::Bool, color::String, alpha::Number)

Plor linear regression and confidence interval on the axis, given

ax Given axis

list_x Array of x

list_y Array of y

intercept Optional: if true, fit the data with an intercept

interval Optional: if true, plot the confidence interval of fitted y

color Color the fitted curve

alpha Transparency of the confidence interval (same color as curve)

The function will filter out the NaN from the given data, make linear regression using function line_regress, sort the results by the value of x from low to high, store the results in a LinearRegressionResult type struct, and plot the regression line on the given axis. Also, if interval is set to true, the function will also plot the confidence intervals of the predicted y.

PlotPlants.line_regress — Function

line_regress(list_x::Array, list_y::Array; intercept::Bool, sorting::Bool)

Make linear regression and return the fitted results, given

list_x Array of x, can be NaN

list_y Array of y, can be NaN

intercept Optional: if true use intercept in the fitting

sorting Optional: if true, fort the values

PlotPlants.LinearRegressionResult — Type

mutable struct LinearRegressionResult

Fields

lm

Regression result

inter

Intercept of fitting

slope

Slope of fitting

inter_p

P value of intercept

slope_p

P value of slopes

inter_ci

Confidence interval of intercept

slope_ci

Confidence interval of slopes

df

Predictions DataFrame

Further, the PlotPlants module provides a quick tool to test whether the slope differs from a given value, see

PlotPlants.line_regress_test_slope — Function

line_regress_test_slope(list_x::Array, list_y::Array; slope::Number, intercept::Bool)

Make linear regression and return the p value of whether the regression slope differs from the given slope, given

list_x Array of x, can be NaN

list_y Array of y, can be NaN

slope Slope to test

intercept Optional: if true use intercept in the fitting

Statistical tools

In many cases, displaying mean absolute (precentage) error make the figure look much better. Thus, the PlotPlants function provides these convenient functions.

PlotPlants.mae — Function

mae(x::Array)

Return the mean absolute error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PlotPlants.mape — Function

mae(x::Array)

Return the mean absolute percentage error by ommiting the NaN, given

y Array of numbers, can be NaN

pred Array of predictions, can be NaN

PlotPlants.nanmean — Function

nanmean(x::Array)

Return the mean of array by ommiting the NaN, given

x Array of numbers, can be NaN

PlotPlants.nanstd — Function

nanstd(x::Array)

Return the std of array by ommiting the NaN, given

x Array of numbers, can be NaN